+++ client/src/App.js
@@ -10,6 +10,7 @@ import Signup from "./pages/Signup";
 import Login from "./pages/Login";
 import Navbar from "./components/navbar/Navbar";
 import Profile from "./pages/Profile";
+import EditProfile from "./pages/EditProfile";
 import Photo from "./pages/Photo";
 import Availability from "./pages/Availability";
 import Payment from "./pages/Payment";
@@ -32,16 +33,29 @@ function App() {
             <Navbar />
             <Route path="/signup" component={Signup} />
             <Route path="/login" component={Login} />
-            <PrivateRoute path="/" exact component={Profile} />
-            <PrivateRoute path="/dashboard/profile" component={Profile} />
-            <PrivateRoute path="/dashboard/photo" component={Photo} />
+            <PrivateRoute exact path="/" component={Profile} />
             <PrivateRoute
+              exact
+              path="/dashboard/editprofile"
+              component={EditProfile}
+            />
+            <PrivateRoute exact path="/dashboard/photo" component={Photo} />
+            <PrivateRoute
+              exact
               path="/dashboard/availability"
               component={Availability}
             />
-            <PrivateRoute path="/dashboard/payment" component={Payment} />
-            <PrivateRoute path="/dashboard/security" component={Security} />
-            <PrivateRoute path="/dashboard/settings" component={Settings} />
+            <PrivateRoute exact path="/dashboard/payment" component={Payment} />
+            <PrivateRoute
+              exact
+              path="/dashboard/security"
+              component={Security}
+            />
+            <PrivateRoute
+              exact
+              path="/dashboard/settings"
+              component={Settings}
+            />
           </AuthDispatchContext.Provider>
         </AuthStateContext.Provider>
       </BrowserRouter>
