+++ client/src/pages/Settings.js
@@ -1,41 +1,82 @@
-import React, { useContext } from "react";
-import { withRouter } from "react-router-dom";
-import { Grid, Dialog, Hidden, Paper } from "@material-ui/core";
-import Menu from "../components/dashboard/Menu";
-import BackButton from "../components/dashboard/BackButton";
-import { UserContext } from "../context/Context";
+import React from "react";
+import { Box, Hidden, Paper, makeStyles } from "@material-ui/core";
+import SettingsMenu from "../components/dashboard/SettingsMenu";
+
+const useStyles = makeStyles(() => ({
+  centerPaper: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+  },
+}));
 
 function Settings() {
-  const { dashboardDialogOpen } = useContext(UserContext);
-
+  const classes = useStyles();
   return (
-    <Grid container style={{ paddingTop: "90px", height: "100vh" }}>
-      <Grid item lg={3} md={3} sm={4} xs={12}>
-        <Menu />
-      </Grid>
-      <Hidden xsDown>
-        <Grid item lg={9} md={9} sm={8} style={{ padding: "15px" }}>
-          <Paper square elevation={3} style={{ width: "100%", height: "100%" }}>
-            Settings page
+    <Box
+      display="flex"
+      flexWrap="nowrap"
+      style={{ paddingTop: "90px", minHeight: "100vh" }}
+      bgcolor="grey"
+    >
+      <Hidden smDown>
+        <Box>
+          <SettingsMenu />
+        </Box>
+      </Hidden>
+      <Hidden smDown>
+        <Box flexGrow={1}>
+          <Paper
+            square
+            elevation={3}
+            className={classes.centerPaper}
+            style={{
+              paddingBottom: "50px",
+              marginBottom: "25px",
+              width: "690px",
+              marginLeft: "auto",
+              marginRight: "auto",
+            }}
+          >
+            {/* Your Component Goes Here */}
           </Paper>
-        </Grid>
+        </Box>
+      </Hidden>
+      <Hidden xsDown mdUp>
+        <Box flexGrow={1}>
+          <Paper
+            square
+            elevation={3}
+            className={classes.centerPaper}
+            style={{
+              paddingBottom: "50px",
+              marginBottom: "25px",
+              width: "580px",
+              marginLeft: "auto",
+              marginRight: "auto",
+            }}
+          >
+            {/* Your Component Goes Here */}
+          </Paper>
+        </Box>
       </Hidden>
       <Hidden smUp>
-        <Dialog open={dashboardDialogOpen} fullScreen>
-          <BackButton />
-          <Grid item lg={8} md={8} sm={8} xs={12} style={{ padding: "15px" }}>
-            <Paper
-              square
-              elevation={3}
-              style={{ width: "100%", height: "100%" }}
-            >
-              Settings mobile
-            </Paper>
-          </Grid>
-        </Dialog>
+        <Box flexGrow={1} style={{}}>
+          <Paper
+            square
+            elevation={3}
+            className={classes.centerPaper}
+            style={{
+              width: "100%",
+              height: "100%",
+            }}
+          >
+            {/* Your Component Goes Here */}
+          </Paper>
+        </Box>
       </Hidden>
-    </Grid>
+    </Box>
   );
 }
 
-export default withRouter(Settings);
+export default Settings;
